Test WebSocketMessageHandler with MessagePackFormatter
